// Source/Client/Interop/JSInteropTypes.Generated.ts
export interface IControlsInterop {
    initialize(containerId: string): Promise<void>;
    dispose(): Promise<void>;
    addButton(caption: string, left: number, top: number): Promise<string>;
    addTextBox(isMultiLine: boolean, left: number, top: number): Promise<string>;
    getButtonCaption(buttonName: string): Promise<string>;
    getTextBoxText(textBoxName: string): Promise<string>;
    hideControl(controlName: string): Promise<void>;
    moveControl(controlName: string, x: number, y: number): Promise<void>;
    removeControl(controlName: string): Promise<void>;
    setButtonCaption(buttonName: string, caption: string): Promise<void>;
    setTextBoxText(textBoxName: string, text: string): Promise<void>;
    setSize(controlName: string, width: number, height: number): Promise<void>;
    showControl(controlName: string): Promise<void>;
    getHeight(): Promise<number>;
    getWidth(): Promise<number>;
}